# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Comment Sentiment Analyzer API
  description: |
    The Comment Sentiment Analyzer API provides comprehensive sentiment analysis for social media comments across multiple platforms. 
    This RESTful API enables content creators to connect their social media accounts, analyze comment sentiment, and gain actionable insights about their audience engagement.

    ## Features
    - Multi-platform social media integration (YouTube, Instagram, Twitter/X, TikTok)
    - AI-powered sentiment analysis with emotion detection
    - Theme clustering and keyword extraction
    - Real-time analysis progress tracking
    - Comprehensive export capabilities
    - Analysis comparison and trending

    ## Authentication
    The API uses JWT (JSON Web Token) based authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Rate Limiting
    API requests are rate limited to ensure fair usage:
    - Authentication: 100 requests/hour
    - Platform Operations: 50 requests/hour  
    - Analysis Operations: 100 requests/hour
    - Export Operations: 20 requests/hour

  version: 1.0.0
  contact:
    name: Comment Sentiment Analyzer Support
    email: support@comment-analyzer.com
    url: https://docs.comment-analyzer.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.comment-analyzer.com/v1
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user account
      description: |
        Creates a new user account with email and password. Returns user data and JWT token for immediate authentication.
        Password must meet security requirements (8+ characters, uppercase, lowercase, number, special character).
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            examples:
              basic_registration:
                summary: Basic user registration
                value:
                  email: "user@example.com"
                  password: "SecurePass123!"
                  firstName: "John"
                  lastName: "Doe"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthResponse"
              examples:
                success:
                  summary: Successful registration
                  value:
                    success: true
                    data:
                      user:
                        id: "user_123"
                        email: "user@example.com"
                        firstName: "John"
                        lastName: "Doe"
                        createdAt: "2024-01-15T10:30:00Z"
                        emailVerified: false
                      token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      expiresAt: "2024-01-22T10:30:00Z"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "429":
          $ref: "#/components/responses/RateLimitError"
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user login
      description: |
        Validates user credentials and returns JWT token for API access. 
        Updates user's last login timestamp and creates cached session.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              user_login:
                summary: User login
                value:
                  email: "user@example.com"
                  password: "SecurePass123!"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "429":
          $ref: "#/components/responses/RateLimitError"
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Returns the authenticated user's profile information
      operationId: getUserProfile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/AuthenticationError"

    put:
      tags:
        - Authentication
      summary: Update user profile
      description: Updates the authenticated user's profile information
      operationId: updateUserProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"

  # Platform Connection Endpoints
...
  /platforms/disconnect/{platform}:
    delete:
      tags:
        - Platforms
      summary: Disconnect social media platform
      description: Removes the connection to a social media platform and revokes stored tokens
      operationId: disconnectPlatform
      security:
        - BearerAuth: []
      parameters:
        - name: platform
          in: path
          required: true
          description: Social media platform to disconnect
          schema:
            $ref: "#/components/schemas/Platform"
      responses:
        "204":
          description: Platform disconnected successfully
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # Posts Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Platform:
      type: string
      enum: [YOUTUBE, INSTAGRAM, TWITTER, TIKTOK]
      description: Supported social media platforms
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Optional message providing additional context
        correlationId:
          type: string
          description: Request correlation ID for debugging
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ApiError:
      type: object
      required:
        - message
        - code
        - status
      properties:
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        status:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error details
        correlationId:
          type: string
          description: Request correlation ID for debugging
        userMessage:
          type: string
          description: User-friendly error message
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Field-specific validation errors
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          description: User's password (must meet security requirements)
          example: "SecurePass123!"
        firstName:
          type: string
          maxLength: 50
          description: User's first name
          example: "John"
        lastName:
          type: string
          maxLength: 50
          description: User's last name
          example: "Doe"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        password:
          type: string
          description: User's password
          example: "SecurePass123!"
    UpdateProfileRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Updated email address
        firstName:
          type: string
          maxLength: 50
          description: Updated first name
        lastName:
          type: string
          maxLength: 50
          description: Updated last name
    AuthResponse:
      type: object
      required:
        - user
        - token
        - expiresAt
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expiresAt:
          type: string
          format: date-time
          description: Token expiration timestamp
    User:
      type: object
      required:
        - id
        - email
        - createdAt
        - updatedAt
        - emailVerified
      properties:
        id:
          type: string
          description: Unique user identifier
          example: "user_123"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        firstName:
          type: string
          description: User's first name
          example: "John"
        lastName:
          type: string
          description: User's last name
          example: "Doe"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last profile update timestamp
        lastLoginAt:
          type: string
          format: date-time
          description: Last login timestamp
        emailVerified:
          type: boolean
          description: Whether email is verified

    # Platform Types
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    $ref: "#/components/schemas/ApiError"
          examples:
            validation_error:
              summary: Validation error example
              value:
                success: false
                error:
                  message: "Validation failed"
                  code: "VALIDATION_ERROR"
                  status: 400
                  fieldErrors:
                    email: ["Email is required", "Email format is invalid"]
                    password: ["Password must be at least 8 characters"]
    AuthenticationError:
      description: Authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    $ref: "#/components/schemas/ApiError"
          examples:
            auth_error:
              summary: Authentication error
              value:
                success: false
                error:
                  message: "Authentication required"
                  code: "AUTHENTICATION_ERROR"
                  status: 401
                  userMessage: "Please log in to continue"
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    $ref: "#/components/schemas/ApiError"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    $ref: "#/components/schemas/ApiError"
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    $ref: "#/components/schemas/ApiError"
      headers:
        X-RateLimit-Limit:
          description: Request limit per time window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when rate limit resets (Unix timestamp)
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
...